FIX: myprofile page 데이터 패칭 병렬 변경
